add ConstMath.Rules, fix fromRational folding
fromRational was being folded too early, before the type was known.  Now, in
order to perform constant folding of fromRational (and you almost certainly
want this), client code needs to import ConstMath.Rules somewhere near the top
of the modules to compile in order to get some specific RULEs in scope.