Add ghc-7.6.1 support
This conditionally handles pretty printing differently, as ghc-7.6
requires that showSDoc takes DynFlags. Hence we lift the pretty function
into CoreM, where we can access the DynFlags. As we don't want to pass
that into all the collapse functions, we match the function name by
constructing it explicitly from its moduleName and occName, which also
ensures that matching is not affected by compiler flags.